@@ -3,13 +3,13 @@ using System.Linq;
 using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
-using Newtonsoft.Json.Linq;
 using System.Net.Http;
 using System.Text;
 using System.Security.Cryptography;
 using System.Net.Http.Headers;
 using System.IO;
 using Microsoft.Extensions.Logging;
+using System.Text.Json;
 
 public class Config
 {
@@ -85,6 +85,17 @@ public class Config
         }
     }
 
+    private class Keys
+    {
+        public Key[] value = null;
+    }
+
+    private class Key
+    {
+        public string id = null;
+        public string value = null;
+    }
+
     private async static Task<(string, string)> GetIdAndSecret()
     {
         try
@@ -97,9 +108,8 @@ public class Config
                 string url = $"https://management.azure.com{AppConfigResourceId}/ListKeys?api-version=2019-02-01-preview";
                 byte[] bytes = client.UploadData(new Uri(url), Encoding.UTF8.GetBytes(string.Empty));
                 string raw = Encoding.UTF8.GetString(bytes);
-                dynamic json = JObject.Parse(raw);
-                JArray values = json.value;
-                dynamic pri = values.First();
+                var json = JsonSerializer.Deserialize<Keys>(raw);
+                var pri = json.value.First();
                 return ((string)pri.id, (string)pri.value);
             }
         }
@@ -114,6 +124,17 @@ public class Config
                 throw;
             }
         }
+    }
+
+    private class Items
+    {
+        public Item[] items = null;
+    }
+
+    private class Item
+    {
+        public string key = null;
+        public string value = null;
     }
 
     public async static Task<Dictionary<string, string>> Load(string[] filters, bool useFullyQualifiedName = false)
@@ -156,8 +177,8 @@ public class Config
                 var raw = await response.Content.ReadAsStringAsync();
 
                 // look for key/value pairs
-                dynamic json = JObject.Parse(raw);
-                foreach (dynamic item in json.items)
+                var json = JsonSerializer.Deserialize<Items>(raw);
+                foreach (var item in json.items)
                 {
                     var key = (useFullyQualifiedName) ? (string)item.key : ((string)item.key).Split(":").Last().ToUpper();
                     var val = (string)item.value;
